pyproject.toml
[project]
name = "stock-signal"
version = "0.1.0"
description = "Medium-horizon equity signal engine for US and India equities"
requires-python = ">=3.13,<3.15"
dependencies = [
    "duckdb>=1.2",
    "pandas>=2.2",
    "numpy>=2.0",
    "httpx>=0.28",
    "pyarrow>=19.0",
    "python-dotenv>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
]

[project.scripts]
stock-signal = "stock_signal.__main__:main"

[tool.ruff]
target-version = "py313"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_signal"]
